+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   for any valid byte string to be a dirname, instead of just valid UTF-8.
 - BREAKING: Can no longer open irregular files as `File`.
 - Fix incorrect checksum updating when initializing directory, leading to the last entry being reported as corrupted.
+- Fix inode size not being updated when writing to block map files
 
 ## 0.1.0-alpha.3